Changed vectors from ints to doubles

+++ trunk/src/pso.c
@@ -3,7 +3,7 @@
 
 double
 get_signal(map *mp, vector *pos) {
-	return mp->signal[pos->x*mp->size->x + pos->y];
+	return mp->signal[(int)pos->x*(int)mp->size->x + (int)pos->y];
 }
 
 void
@@ -18,5 +18,5 @@ next_iteration(particle *ptcl, double w, double c1, double c2, double r1, double
 void save_state(particle *particles, int p, FILE *out) {
 	int i;
 	for (i=0;i<p;i++)
-		fprintf(out, "Particle %d: %d, %d\n", i+1, particles -> position -> x, particles -> position -> y);
+		fprintf(out, "Particle %d: %lf, %lf\n", i+1, particles -> position -> x, particles -> position -> y);
 }
